editable links + optimized easyEdit in components
