+++ src/pages/settings/profile/index.css
@@ -2,32 +2,30 @@
 @tailwind components;
 @tailwind utilities;
 
-.field-name{
-  @apply  rounded-md text-lg md:text-xl font-medium text-indigo-900;
+.field-name {
+  @apply rounded-md text-lg md:text-xl font-medium text-indigo-900;
 }
-.field-name::after{
-  content: ""
+.field-name::after {
+  content: '';
 }
 @media (min-width: 768px) {
-  .field-name::after{
-    content: ":"
+  .field-name::after {
+    content: ':';
   }
 }
-.no-after::after{
-  content: ""
-
+.no-after::after {
+  content: '';
 }
-.field-value{
+.field-value {
   @apply bg-indigo-100 p-3 rounded-md mb-4 sm:mb-0;
 }
-h1{
+h1 {
   @apply font-bold text-2xl md:text-3xl text-indigo-950 capitalize;
 }
 
-.field-link-name{
+.field-link-name {
   @apply pl-3 rounded-md text-xl font-medium text-indigo-900;
 }
-.field-link-value{
+.field-link-value {
   @apply p-3 rounded-md border-2 flex flex-row flex-wrap items-center gap-4 mb-4 sm:mb-0;
-  
 }
